requirements.txt
streamlit>=1.43,<2
openai>=1.68.2,<2
pillow>=11.1,<12
pandas>=2.2,<3
requests>=2.32,<3
beautifulsoup4>=4.12,<5
python-dotenv>=1.0,<2
